Refactor CircuitStructure and CircuitSetter into helpers
